Fix queries test and sonarcloud issue
