+++ faros-airbyte-common/src/github/queries.ts
@@ -1,10 +1,10 @@
 import fs from 'fs-extra';
 import path from 'path';
 
+import {COMMIT_FIELDS_FRAGMENT} from './fragments';
+
 // GraphQL query used to get pull requests
 export const PULL_REQUESTS_QUERY = loadQuery('pull-requests-query.gql');
-
-export const FILES_FRAGMENT = loadFragment('files.gql');
 
 // GraphQL query used to get pull request files
 export const PULL_REQUEST_FILES_QUERY = loadQuery(
@@ -19,17 +19,16 @@ export const ORG_MEMBERS_QUERY = loadQuery('list-members-query.gql');
 
 // GraphQL query used to get commits from repository
 export const COMMITS_QUERY =
-  loadFragment('commit-fields.gql') + loadQuery('commits-query.gql');
+  COMMIT_FIELDS_FRAGMENT + loadQuery('commits-query.gql');
 
 // GraphQL query used to get commits from repository with changedFilesIfAvailable
 export const COMMITS_CHANGED_FILES_IF_AVAILABLE_QUERY =
-  loadFragment('commit-fields.gql') +
+  COMMIT_FIELDS_FRAGMENT +
   loadQuery('commits-changed-files-if-available-query.gql');
 
 // GraphQL query used to get commits from repository with changedFiles
 export const COMMITS_CHANGED_FILES_QUERY =
-  loadFragment('commit-fields.gql') +
-  loadQuery('commits-changed-files-query.gql');
+  COMMIT_FIELDS_FRAGMENT + loadQuery('commits-changed-files-query.gql');
 
 /**
  * Load query file from resources
@@ -39,21 +38,6 @@ export const COMMITS_CHANGED_FILES_QUERY =
 function loadQuery(query: string): string {
   return fs.readFileSync(
     path.join(__dirname, '..', '..', 'resources', 'github', 'queries', query),
-    'utf8'
-  );
-}
-
-function loadFragment(fragment: string): string {
-  return fs.readFileSync(
-    path.join(
-      __dirname,
-      '..',
-      '..',
-      'resources',
-      'github',
-      'fragments',
-      fragment
-    ),
     'utf8'
   );
 }
